requirements.txt
asyncpg
fastapi
uvicorn
python-dotenv
python-multipart
pydantic
pydantic-settings
passlib
bcrypt
email-validator
SQLAlchemy
psycopg2-binary
psycopg-pool
pymongo
langchain
langchain-community
langchain_text_splitters
langgraph-checkpoint-postgres
langchain-pinecone
langchain-huggingface
langchain-groq
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
pinecone-client
sentence-transformers
requests
gunicorn